requirements.txt
# Core dependencies
PyYAML
numpy
pandas
matplotlib
